fix : Redis Config 관련 순환 참조 에러 수정
